make keyboard layout a global
