+++ package.json
@@ -35,6 +35,7 @@
     "build:disposable": "cd src/disposable && tsc",
     "build:dragdrop": "cd src/dragdrop && tsc",
     "build:json": "cd src/json && tsc",
+    "build:keyboard": "cd src/keyboard && tsc",
     "build:messaging": "cd src/messaging && tsc",
     "build:mime": "cd src/mime && tsc",
     "build:platform": "cd src/platform && tsc",
@@ -49,6 +50,7 @@
     "clean:disposable": "rimraf src/disposable/lib",
     "clean:dragdrop": "rimraf src/dragdrop/lib",
     "clean:json": "rimraf src/json/lib",
+    "clean:keyboard": "rimraf src/keyboard/lib",
     "clean:messaging": "rimraf src/messaging/lib",
     "clean:mime": "rimraf src/mime/lib",
     "clean:platform": "rimraf src/platform/lib",
