+++ src/keyboard/package.json
@@ -1,0 +1,31 @@
+{
+  "private": true,
+  "name": "@phosphor/keyboard",
+  "version": "1.0.0",
+  "description": "PhosphorJS - Keyboard",
+  "main": "lib/index.js",
+  "types": "lib/index.d.ts",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/phosphorjs/phosphor.git"
+  },
+  "files": [
+    "lib/*.d.ts",
+    "lib/*.js"
+  ],
+  "directories": {
+    "lib": "lib/"
+  },
+  "author": "S. Chris Colbert <sccolbert@gmail.com>",
+  "contributors": [
+    "A. Darian <git@darian.af>",
+    "Dave Willmer <dave.willmer@gmail.com>",
+    "S. Chris Colbert <sccolbert@gmail.com>",
+    "Steven Silvester <steven.silvester@gmail.com>"
+  ],
+  "license": "BSD-3-Clause",
+  "bugs": {
+    "url": "https://github.com/phosphorjs/phosphor/issues"
+  },
+  "homepage": "https://github.com/phosphorjs/phosphor"
+}
